+++ HISTORY.rst
@@ -7,6 +7,8 @@ Release History
 +++++++++++++++++++++
 
 - Fixed import error in async receiver.
+- Exposed sender/receiver destroy function.
+- Moved receiver.open on_message_received argument to constructor.
 
 
 0.1.0b5 (2018-04-27)